requirements.txt
pypdf2 # PDFs
langchain # Various helpers
chromadb # Vector storage
langchain-community # Helpers
langchain-ollama
ollama # local models
google-genai # for gemini sdk`